Revert "fixes"
This reverts commit 8b14fa2e38e18c1438ebe9233fab28cce7eb044e.

+++ src/eghudm.c
@@ -42,9 +42,7 @@ int FAR CDECL hudSine(int angle) { return nsine(angle); }
 
 /* pitch -> ladder pixel offset: (|pitch|>>6) * 360, taking the high bytes. */
 int FAR CDECL hudPitchScale(int ap) {
-    /* The original works on the raw 16-bit pitch word; negative host ints must
-     * wrap as unsigned DOS words before scaling. */
-    return (int)(((unsigned long)(uint16)ap * 360u) >> 8);
+    return (int)(((unsigned long)(unsigned)ap * 360u) >> 8);
 }
 
 /* ===== drawClipLineGlobal (clipLineFar) =====
